Updated all logic to adder to requirements
The major part of this work was on the forecastSensor
and more precisely on the willPrecipitate method.
The forecastSensor was redone with a timetable of precipitations
for 7 days.
The controller's logic was totally redone from scratch to be
in line with the requirements.
The moisture sensor is now locked between -50 and 150.
And a lot more changes.